chore: split `noirc_frontend/src/tests.rs` into submodules
